+++ config/application.rb
@@ -26,6 +26,10 @@ module Conzealand
     # Initialize configuration defaults for originally generated Rails version.
     config.load_defaults 5.1
 
+    # Disable asset pipeline, should all be moved to webpacker now
+    config.assets.enabled = false
+    config.generators { |g| g.assets false }
+
     # Application configuration can go into files in config/initializers
     # -- all .rb files in that directory are automatically loaded after loading
     # the framework and any gems in your application.